pyproject.toml
[tool.poetry]
name = "darkstain"
version = "0.1.0"
description = ""
authors = ["pzaloznyi <pzaloznyi@gmail.com>"]
readme = "README.md"
packages = [
    { include = "api" },
    { include = "config" }
]

[tool.poetry.dependencies]
python = "^3.11.9"
Django = "^5.1.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
